remove the Client.Connect method, and instead perform that operation during initialization.
It's no longer necessary to call Connect() on the return of a NewClient() call.  Skipping that call was previously allowed since asking for Stdout or Stdin would establish the connection anyway, but this was ridiculous because asking for them both from different goroutines would cause a data race in connection establishment, and there's just no reason for that complexity.